+++ src/MicroElements.Metadata/Metadata/ParserProviderExtensions.cs
@@ -40,7 +40,7 @@ namespace MicroElements.Metadata
         /// <param name="parserProvider"><see cref="IParserProvider"/>.</param>
         /// <param name="sourceRow">Source data.</param>
         /// <returns><see cref="IPropertyValue"/> list.</returns>
-        public static IReadOnlyList<ParseResult<IPropertyValue>> ParsePropertiesAsParseResults(this IParserProvider parserProvider, IReadOnlyDictionary<string, string> sourceRow)
+        public static IReadOnlyList<ParseResult<IPropertyValue>> ParsePropertiesAsParseResults(this IParserProvider parserProvider, IReadOnlyDictionary<string, string?> sourceRow)
         {
             parserProvider.AssertArgumentNotNull(nameof(parserProvider));
             sourceRow.AssertArgumentNotNull(nameof(sourceRow));